Repositórios, Classes Especialistas e Serviços

+++ src/TesteTecnico.ApplicationCore/Entity/TypePerson.cs
@@ -2,7 +2,7 @@
 {
     public class TypePerson
     {
-        public int TypePersonId { get; private set; }
+        public byte TypePersonId { get; private set; }
         public string Description { get; private set; }
 
         public TypePerson() { }
